fix: complete legacy candidate analysis
